+++ src/psd_tools/composer/__init__.py
@@ -187,7 +187,10 @@ def compose_layer(layer, force=False, **kwargs):
             clip_image = compose(
                 layer.clip_layers, bbox=bbox, context=image.copy()
             )
-            mask = image.getchannel('A')
+            if image.mode.endswith('A'):
+                mask = image.getchannel('A')
+            else:
+                mask = Image.new('L', image.size, 255)
             if clip_image.mode.endswith('A'):
                 mask = ImageChops.darker(clip_image.getchannel('A'), mask)
             clip_image.putalpha(mask)
@@ -332,6 +335,7 @@ def apply_effect(layer, backdrop, base_image):
             else:
                 alpha = base_image.convert('L')
             alpha.info['offset'] = base_image.info['offset']
+            flat = alpha.getextrema()[0] < 255
 
             # Expand the image size
             setting = effect.value
@@ -345,7 +349,7 @@ def apply_effect(layer, backdrop, base_image):
 
             if not layer.has_vector_mask() and setting.get(
                 Key.Style
-            ).enum == Enum.InsetFrame:
+            ).enum == Enum.InsetFrame and flat:
                 image = create_stroke_effect(alpha, setting, layer._psd, True)
                 backdrop.paste(image)
             else: