+++ src/psd_tools/composer/vector.py
@@ -101,6 +101,9 @@ def _draw_subpath(subpath, width, height):
     from PIL import Image
     import aggdraw
     mask = Image.new('L', (width, height), 0)
+    if len(subpath) <= 1:
+        logger.warning('not enough knots: %d' % len(subpath))
+        return mask
     path = ' '.join(map(str, _generate_symbol(subpath, width, height)))
     draw = aggdraw.Draw(mask)
     brush = aggdraw.Brush(255)
@@ -156,7 +159,7 @@ def draw_solid_color_fill(size, setting):
     from PIL import Image, ImageDraw, ImageChops
     color = setting.get(Key.Color)
     mode = _COLORSPACE.get(color.classID)
-    fill = tuple(int(x) for x in color.values())
+    fill = tuple(int(x) for x in list(color.values())[:len(mode)])
     canvas = Image.new(mode, size)
     draw = ImageDraw.Draw(canvas)
     draw.rectangle((0, 0, canvas.width, canvas.height), fill=fill)
@@ -230,6 +233,10 @@ def draw_gradient_fill(size, setting):
         Z = _make_reflected_gradient(X, Y, angle)
     elif gradient_kind == Enum.Diamond:
         Z = _make_diamond_gradient(X, Y, angle)
+    elif gradient_kind == b'shapeburst':
+        # Only available in stroke effect.
+        logger.warning('Gradient style not supported: %s' % gradient_kind)
+        Z = np.ones((size[1], size[0])) * 0.5
     else:
         logger.warning('Unknown gradient style: %s.' % (gradient_kind))
         Z = np.ones((size[1], size[0])) * 0.5
@@ -341,7 +348,8 @@ def _apply_color_map(grad, Z):
             mode = _COLORSPACE.get(stop.get(Key.Color).classID)
             s = scalar.get(mode, 1.0)
             location = int(stop.get(Key.Location)) / 4096.
-            color = tuple(s * x for x in stop.get(Key.Color).values())
+            color = list(stop.get(Key.Color).values())[:len(mode)]
+            color = tuple(s * int(x) for x in color)
             if len(X) and X[-1] == location:
                 logger.debug('Duplicate stop at %d' % location)
                 X.pop(), Y.pop()